FIX massaged_url generates key from original target URI even when original_url is overridden to change target URI.
Use original_url instead of record.url for generating key. This has a side_effect of both messaged_url and original_url being in the same UTF-8 encoding. I assume this is the right thing to do: if encoding does not much, timestamp redirect will fail to find the capture.