fix: cdx_writer falls in an infinite loop on a corrupted WARC
if request record (or any non-cdx-generating record) is corrupted,
ArchiveRecordReader raises an error during a iteration for the next record.
If this happens at the last record, cdx_writer falls in a infinite error loop
trying to read gzip header at EOF. If it happens in the middle of the file,
it can discard a healthy record as corrupted. Two counter measures:
- read off record content before skipping to the next cord to detect errors
  and position file pointer correctly within the cycle
- reset() communicates (unexpected) EOF situation to __next__() to prevent
  infinite loop.